requirements.txt
aiogram~=2.11.2
aiohttp~=3.7.4.post0
cachetools~=4.2.0
escpos~=1.8
google-auth-oauthlib~=0.4.3
gspread-asyncio~=1.4.0
phonenumbers~=8.12.19
pygsheets~=2.0.5
requests~=2.25.1
schedule~=1.0.0
SQLAlchemy-Utils~=0.36.8
SQLAlchemy~=1.3.22
